fixed a bug

+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/entity/ContractSupplierDetail.java
@@ -24,7 +24,7 @@ public class ContractSupplierDetail implements BaseEntity
     private Integer boughtAmount;
 
     @Column(name = "price_per_unit")
-    private Integer pricePerUnit;
+    private Float pricePerUnit;
 
     @ManyToOne
     @JoinColumn(name = "contract_supplier_id")
